Rename SqlText to SqlFormatting.
